+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/OutpostNetherStructure.java
@@ -21,7 +21,7 @@ import net.minecraft.world.gen.feature.StructurePoolFeatureConfig;
 import java.util.List;
 
 
-public class OutpostNetherStructure extends AbstractBaseStructure {
+public class OutpostNetherStructure extends AbstractBaseStructure<DefaultFeatureConfig> {
     private final Identifier START_POOL;
     private static final List<SpawnSettings.SpawnEntry> MONSTER_SPAWNS = Lists.newArrayList(new SpawnSettings.SpawnEntry(EntityType.PIGLIN, 10, 1, 1));